+++ sdk/authorization/azure-resourcemanager-authorization/src/main/java/com/azure/resourcemanager/authorization/fluent/models/MicrosoftGraphStatus.java
@@ -8,7 +8,7 @@ import com.azure.core.util.ExpandableStringEnum;
 import java.util.Collection;
 
 /**
- * status.
+ * Represents a status.
  */
 public final class MicrosoftGraphStatus extends ExpandableStringEnum<MicrosoftGraphStatus> {
     /**