+++ sdk/monitor/azure-resourcemanager-monitor/src/main/java/com/azure/resourcemanager/monitor/implementation/VMInsightsClientImpl.java
@@ -55,7 +55,7 @@ public final class VMInsightsClientImpl implements VMInsightsClient {
      * REST calls.
      */
     @Host("{$host}")
-    @ServiceInterface(name = "MonitorClientVMInsig")
+    @ServiceInterface(name = "MonitorClientVMInsights")
     public interface VMInsightsService {
         @Headers({ "Content-Type: application/json" })
         @Get("/{resourceUri}/providers/Microsoft.Insights/vmInsightsOnboardingStatuses/default")